Add empty objects
As _parent used to be the class id of the parent, it was easy to use a
class already handle by another tool as a parent, by setting the class
id to whatever the user wanted. Now, _parent has to be an object
representing a HTB class (for now). This commit allow to restore this
"feature".