Add a way to not change the define way of classes
Before, rules classes could be defined without redefining an __init__
function during the inheritance. With rate and ceil transformed in
properties, it was not possible.

I choose the solution of re-defining the rate and ceil attributes as
properties during the __init__ of EmptyHTBClass. It has the advantage to
not change **at all** the way of declaring the rules.